Updated optional_rules to AppV2
