+++ betterrolls-swade2/templates/optional_rules.hbs
@@ -1,4 +1,4 @@
-<form id="brsw-optional-rules">
+<div>
     <h2>{{ localize "BRSW.OptionalRules" }}</h2>
     <table>
         {{# each rules }}<tr>
@@ -22,4 +22,4 @@
         </tr>
     </table>
     <button type="submit"><i class="far fa-save"></i>{{localize "BRSW.Save"}}</button>
-</form>
+</div>
